requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
redis==5.0.7
python-dotenv==1.0.1
